extracted common code
