+++ client/src/components/Nav/UserMenu.tsx
@@ -19,25 +19,15 @@ export default function UserMenu (): JSX.Element {
 
   return (
     <div>
-      <IconButton
-        size="large"
-        onClick={handleMenu}
-        color="inherit"
-      >
+      <IconButton size="large" onClick={handleMenu} color="inherit">
         <AccountCircle/>
       </IconButton>
       <Menu
         id="menu-appbar"
         anchorEl={anchorEl}
-        anchorOrigin={{
-          vertical: 'top',
-          horizontal: 'right'
-        }}
+        anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
         keepMounted
-        transformOrigin={{
-          vertical: 'top',
-          horizontal: 'right'
-        }}
+        transformOrigin={{ vertical: 'top', horizontal: 'right' }}
         open={Boolean(anchorEl)}
         onClose={handleClose}
       >
