+++ client/src/components/Signup/index.tsx
@@ -1,32 +1,19 @@
 import * as React from 'react'
 import Box from '@mui/material/Box'
-import { Tab, Tabs } from '@mui/material'
 import BusinessSignup from './BusinessSignup'
-import TabPanel from '../Shared/TabPanel'
 import Container from '@mui/material/Container'
-import InfluencerSignup from './InfluencerSignup'
 import { containerStyle } from '../../utils/utils'
+import AuthTabs from '../Shared/AuthTabs'
+import InfluencerAuthTab from '../Shared/InfluencerAuthTab'
 
 export default function Signup (): JSX.Element {
-  const [value, setValue] = React.useState(0)
-
-  const handleChange = (event: React.SyntheticEvent, newValue: number): void => {
-    setValue(newValue)
-  }
-
   return (
     <Container component="main" maxWidth="xs">
       <Box sx={{ ...containerStyle.centeredBox, marginTop: 8 }}>
-        <Tabs value={value} onChange={handleChange}>
-          <Tab label="Business"/>
-          <Tab label="Influencer"/>
-        </Tabs>
-        <TabPanel value={value} index={0}>
-          <BusinessSignup/>
-        </TabPanel>
-        <TabPanel index={value} value={1}>
-          <InfluencerSignup/>
-        </TabPanel>
+        <AuthTabs
+          businessComp={<BusinessSignup/>}
+          influencerComp={<InfluencerAuthTab label="Sync with Google"/>}
+        />
       </Box>
     </Container>
   )
